refactor(canada): declare CANADA_ROAD_SOURCES in the importable core
tests/canada-roads.test.mts carried its own copy of the four descriptors
because src/services/canada-roads.ts reaches Vite-side imports and cannot be
loaded from a test. A cloned constant cannot fail when the real one changes,
so no assertion about the descriptors was load-bearing.

Move the declaration to canada-roads-core.ts, which has no imports at all, and
re-export it so the public path is unchanged. Both tests now read the real
list — the freshness test by import rather than by scraping source text.
