+++ src/services/canada-roads-core.ts
@@ -33,6 +33,21 @@ export interface CanadaRoadSourceDescriptor {
   onDemand: boolean;
 }
 
+/**
+ * The four road feeds that union onto the `canadaRoads` map layer.
+ *
+ * Declared here rather than in `canada-roads.ts` so tests can assert against the
+ * real descriptors: that module reaches Vite-side imports (`import.meta.env`),
+ * so a test importing it has to fall back to reading source text or cloning this
+ * list — and a cloned copy cannot fail when the real one changes.
+ */
+export const CANADA_ROAD_SOURCES: readonly CanadaRoadSourceDescriptor[] = Object.freeze([
+  { key: 'canadaRoads', source: 'ontario-511', jurisdiction: 'ON', onDemand: false },
+  { key: 'albertaRoads', source: 'alberta-511', jurisdiction: 'AB', onDemand: false },
+  { key: 'torontoRoads', source: 'toronto-roads', jurisdiction: 'Toronto', onDemand: true },
+  { key: 'bcOpen511', source: 'bc-open511', jurisdiction: 'BC', onDemand: true },
+]);
+
 export type CanadaRoadSourceStates = Record<string, CanadaRoadSourceState>;
 
 export function hasHealthyCanadaRoadSource(states: CanadaRoadSourceStates): boolean {
